# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CryptoMarket API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  username: string;
  email: string;
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  email: string;
  role: string;
  /** @nullable */
  avatarUrl?: string | null;
  isActive?: boolean;
  createdAt: string;
  /** @nullable */
  totalPurchases?: number | null;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  description: string;
  /** @nullable */
  shortDescription?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  tags?: string[];
  previewImages?: string[];
  /** @nullable */
  demoUrl?: string | null;
  /** @nullable */
  downloadUrl?: string | null;
  /** @nullable */
  version?: string | null;
  /** @nullable */
  changelog?: string | null;
  /** @nullable */
  documentation?: string | null;
  /** @nullable */
  videoPreviewUrl?: string | null;
  status: string;
  isFeatured?: boolean;
  salesCount?: number;
  /** @nullable */
  rating?: number | null;
  reviewCount?: number;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  description: string;
  shortDescription?: string;
  price: number;
  originalPrice?: number;
  categoryId: number;
  tags?: string[];
  previewImages?: string[];
  demoUrl?: string;
  downloadUrl?: string;
  version?: string;
  documentation?: string;
  videoPreviewUrl?: string;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  shortDescription?: string;
  price?: number;
  originalPrice?: number;
  categoryId?: number;
  tags?: string[];
  previewImages?: string[];
  demoUrl?: string;
  downloadUrl?: string;
  version?: string;
  changelog?: string;
  documentation?: string;
  videoPreviewUrl?: string;
  isFeatured?: boolean;
  status?: string;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  iconName?: string | null;
  productCount?: number;
}

export interface CategoryInput {
  name: string;
  slug: string;
  description?: string;
  iconName?: string;
}

export interface Payment {
  id: number;
  orderId: number;
  amount: number;
  chain: string;
  txHash: string;
  /** @nullable */
  screenshotUrl?: string | null;
  status: string;
  createdAt: string;
}

export interface Order {
  id: number;
  userId: number;
  productId: number;
  /** @nullable */
  productName?: string | null;
  /** @nullable */
  productImage?: string | null;
  amount: number;
  status: string;
  /** @nullable */
  adminNotes?: string | null;
  createdAt: string;
  updatedAt?: string;
  payment?: Payment;
}

export interface OrderInput {
  productId: number;
}

export interface OrderStatusUpdate {
  status: string;
  adminNotes?: string;
}

export interface PaymentSubmission {
  orderId: number;
  chain: string;
  txHash: string;
  amount: number;
  screenshotUrl?: string;
}

export interface PaymentVerification {
  status: string;
  notes?: string;
}

export interface Wallet {
  id: number;
  chain: string;
  address: string;
  /** @nullable */
  label?: string | null;
  isActive: boolean;
}

export interface WalletAddresses {
  wallets: Wallet[];
}

export interface WalletInput {
  chain: string;
  address: string;
  label?: string;
  isActive?: boolean;
}

export interface Review {
  id: number;
  productId: number;
  userId: number;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface ReviewInput {
  rating: number;
  comment?: string;
}

export interface Conversation {
  id: number;
  userId: number;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  lastMessage?: string | null;
  /** @nullable */
  lastMessageAt?: string | null;
  unreadCount?: number;
  createdAt: string;
}

export interface ConversationInput {
  subject: string;
}

export interface Message {
  id: number;
  conversationId: number;
  senderId: number;
  /** @nullable */
  senderName?: string | null;
  /** @nullable */
  senderAvatar?: string | null;
  content: string;
  /** @nullable */
  fileUrl?: string | null;
  isRead?: boolean;
  createdAt: string;
}

export interface MessageInput {
  content: string;
  fileUrl?: string;
}

export interface Notification {
  id: number;
  userId: number;
  type: string;
  title: string;
  message: string;
  /** @nullable */
  link?: string | null;
  isRead: boolean;
  createdAt: string;
}

export interface Download {
  id: number;
  orderId: number;
  productId: number;
  productName: string;
  /** @nullable */
  productImage?: string | null;
  downloadedAt: string;
  /** @nullable */
  version?: string | null;
}

export interface DownloadLink {
  url: string;
  expiresAt: string;
}

export interface Ticket {
  id: number;
  userId: number;
  /** @nullable */
  username?: string | null;
  subject: string;
  /** @nullable */
  description?: string | null;
  status: string;
  priority?: string;
  /** @nullable */
  adminReply?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface TicketInput {
  subject: string;
  description: string;
  priority?: string;
}

export interface TicketUpdate {
  status?: string;
  adminReply?: string;
  priority?: string;
}

export interface CategorySales {
  category: string;
  sales: number;
  revenue: number;
}

export interface AdminAnalytics {
  totalRevenue: number;
  totalOrders: number;
  totalUsers: number;
  totalProducts: number;
  pendingOrders: number;
  pendingPayments?: number;
  recentOrders: Order[];
  topProducts: Product[];
  salesByCategory: CategorySales[];
}

export interface RevenueDataPoint {
  date: string;
  revenue: number;
  orders: number;
}

export interface RevenueAnalytics {
  daily: RevenueDataPoint[];
  monthly: RevenueDataPoint[];
  total: number;
}

export interface AdminUserUpdate {
  isActive?: boolean;
  role?: string;
}

export type ListProductsParams = {
page?: number;
limit?: number;
category?: string;
search?: string;
sort?: string;
minPrice?: number;
maxPrice?: number;
};