# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CryptoMarket API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ORDERS
...
  /downloads/{orderId}:
    get:
      operationId: getDownloadLink
      tags: [downloads]
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Download link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadLink"

  # SUPPORT TICKETS
...
  /admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteAdminUser
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/products/{id}/approve:
    patch:
      operationId: approveProduct
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # AUTH
    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        totalPurchases:
          type: ["integer", "null"]

    # PRODUCTS
    Product:
      type: object
      required: [id, name, slug, description, price, categoryId, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        shortDescription:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        previewImages:
          type: array
          items:
            type: string
        demoUrl:
          type: ["string", "null"]
        downloadUrl:
          type: ["string", "null"]
        version:
          type: ["string", "null"]
        changelog:
          type: ["string", "null"]
        documentation:
          type: ["string", "null"]
        videoPreviewUrl:
          type: ["string", "null"]
        status:
          type: string
        isFeatured:
          type: boolean
        salesCount:
          type: integer
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
...
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        iconName:
          type: ["string", "null"]
        productCount:
          type: integer
    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        iconName:
          type: string

    # ORDERS
...
    DownloadLink:
      type: object
      required: [url, expiresAt]
      properties:
        url:
          type: string
        expiresAt:
          type: string

    # TICKETS
...
    AdminUserUpdate:
      type: object
      properties:
        isActive:
          type: boolean
        role:
          type: string